python-table-extractor/requirements.txt
Flask==2.3.3
camelot-py[cv]==0.10.1
requests==2.31.0
pandas==2.3.2
tabulate==0.9.0
